pyproject.toml
[tool.poetry]
name = "pixel-battle"
version = "0.1.0"
description = ""
authors = ["Renat Yakublevich <merlinincorp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
redis = "^5.0.7"
pydantic-settings = "^2.3.4"
ruff = "^0.5.2"
uvicorn = "^0.30.1"
fastapi = "^0.111.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
